Readequando a arquitetura MVC.
