--- backend/src/main/java/com/auto_mendes/backend/dto/response/EmployeeResponseDTO.java
@@ -1,14 +1,0 @@
-package com.auto_mendes.backend.dto.response;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-public record EmployeeResponseDTO(
-		String id,
-		String name,
-		String email,
-		String matriculation, 
-		String phone,
-		LocalDate birthDate,
-		BigDecimal commission
-) {}
