--- backend/src/main/java/com/auto_mendes/backend/dto/response/ExceptionResponseDTO.java
@@ -1,13 +1,0 @@
-package com.auto_mendes.backend.dto.response;
-
-import java.time.LocalDateTime;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-public record ExceptionResponseDTO(
-		@JsonFormat(pattern = "yyyy-MM-dd HH:ss")
-		LocalDateTime localDateTime,
-		Integer status,
-		String message,
-		String path
-) {}
