--- backend/src/main/java/com/auto_mendes/backend/enums/EmployeeType.java
@@ -1,5 +1,0 @@
-package com.auto_mendes.backend.enums;
-
-public enum EmployeeType {
-	MANAGER, ASSISTANT_MANAGER, SALER;
-}
