--- backend/src/main/java/com/auto_mendes/backend/enums/PaymentType.java
@@ -1,5 +1,0 @@
-package com.auto_mendes.backend.enums;
-
-public enum PaymentType {
-	CASH, CREDIT_CARD, DEBIT_CARD, PIX, BANK_TRANSFER;
-}
