--- backend/src/main/java/com/auto_mendes/backend/enums/TransmissionType.java
@@ -1,5 +1,0 @@
-package com.auto_mendes.backend.enums;
-
-public enum TransmissionType {
-	AUTO, MANUAL;
-}
