+++ backend/src/main/java/com/auto_mendes/backend/mapper/EmployeeMapper.java
@@ -2,12 +2,12 @@ package com.auto_mendes.backend.mapper;
 
 import org.mapstruct.Mapper;
 
-import com.auto_mendes.backend.dto.request.EmployeeRequestDTO;
-import com.auto_mendes.backend.dto.response.EmployeeResponseDTO;
-import com.auto_mendes.backend.persistence.entity.AssistantManager;
-import com.auto_mendes.backend.persistence.entity.Employee;
-import com.auto_mendes.backend.persistence.entity.Manager;
-import com.auto_mendes.backend.persistence.entity.Saler;
+import com.auto_mendes.backend.model.dto.request.EmployeeRequestDTO;
+import com.auto_mendes.backend.model.dto.response.EmployeeResponseDTO;
+import com.auto_mendes.backend.model.entity.AssistantManager;
+import com.auto_mendes.backend.model.entity.Employee;
+import com.auto_mendes.backend.model.entity.Manager;
+import com.auto_mendes.backend.model.entity.Saler;
 
 @Mapper
 public interface EmployeeMapper {
