+++ backend/src/main/java/com/auto_mendes/backend/model/entity/Manager.java
@@ -1,0 +1,18 @@
+package com.auto_mendes.backend.model.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.PrimaryKeyJoinColumn;
+import jakarta.persistence.Table;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@Entity
+@Table(name = "manager_tb")
+@PrimaryKeyJoinColumn(name = "id")
+public class Manager extends Employee {
+	private static final long serialVersionUID = 1L;
+}
