+++ backend/src/main/java/com/auto_mendes/backend/model/entity/Model.java
@@ -1,0 +1,37 @@
+package com.auto_mendes.backend.model.entity;
+
+import java.io.Serializable;
+import java.util.List;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "model_tb")
+public class Model implements Serializable {
+	private static final long serialVersionUID = 1L;
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.UUID)
+	private String id;
+	@Column(nullable = false, unique = true, length = 50)
+	private String name;
+	@ManyToOne
+	@JoinColumn(name = "brand_id")
+	private Brand brand;
+	@OneToMany(mappedBy = "model")
+	private List<Vehicle> vehicles;
+}
