+++ backend/src/main/java/com/auto_mendes/backend/model/entity/SaleVehicleId.java
@@ -1,0 +1,15 @@
+package com.auto_mendes.backend.model.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Embeddable
+public class SaleVehicleId {
+	private String saleId;
+	private String vehicleId;
+}
