+++ backend/src/main/java/com/auto_mendes/backend/model/enums/PaymentType.java
@@ -1,0 +1,5 @@
+package com.auto_mendes.backend.model.enums;
+
+public enum PaymentType {
+	CASH, CREDIT_CARD, DEBIT_CARD, PIX, BANK_TRANSFER;
+}
