+++ backend/src/main/java/com/auto_mendes/backend/model/enums/TransmissionType.java
@@ -1,0 +1,5 @@
+package com.auto_mendes.backend.model.enums;
+
+public enum TransmissionType {
+	AUTO, MANUAL;
+}
