--- backend/src/main/java/com/auto_mendes/backend/persistence/entity/Brand.java
@@ -1,32 +1,0 @@
-package com.auto_mendes.backend.persistence.entity;
-
-import java.io.Serializable;
-import java.util.List;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "brand_tb")
-public class Brand implements Serializable {
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.UUID)
-	private String id;
-	@Column(nullable = false, unique = true, length = 50)
-	private String name;
-	@OneToMany(mappedBy = "brand")
-	private List<Model> models;
-}
