--- backend/src/main/java/com/auto_mendes/backend/persistence/entity/SaleVehicleId.java
@@ -1,15 +1,0 @@
-package com.auto_mendes.backend.persistence.entity;
-
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Embeddable
-public class SaleVehicleId {
-	private String saleId;
-	private String vehicleId;
-}
