--- backend/src/main/java/com/auto_mendes/backend/persistence/entity/Saler.java
@@ -1,28 +1,0 @@
-package com.auto_mendes.backend.persistence.entity;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.PrimaryKeyJoinColumn;
-import jakarta.persistence.Table;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-@Entity
-@Table(name = "saler_tb")
-@PrimaryKeyJoinColumn(name = "id")
-public class Saler extends Employee {
-	private static final long serialVersionUID = 1L;
-	
-	@Column(nullable = false, precision = 20, scale = 2)
-	private BigDecimal commission;
-	@OneToMany(mappedBy = "saler")
-	private List<Sale> sales;
-}
