+++ backend/src/main/java/com/auto_mendes/backend/repository/AssistantManagerRepository.java
@@ -1,0 +1,11 @@
+package com.auto_mendes.backend.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.auto_mendes.backend.model.entity.AssistantManager;
+
+@Repository
+public interface AssistantManagerRepository extends JpaRepository<AssistantManager, String> {
+	boolean existsByEmailOrMatriculationOrPhone(String email, String matriculation, String phone);
+}
