+++ backend/src/main/java/com/auto_mendes/backend/repository/EmployeeRepository.java
@@ -1,0 +1,10 @@
+package com.auto_mendes.backend.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.auto_mendes.backend.model.entity.Employee;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, String> {
+}
