+++ backend/src/main/java/com/auto_mendes/backend/service/EmployeeService.java
@@ -3,9 +3,9 @@ package com.auto_mendes.backend.service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.auto_mendes.backend.dto.request.EmployeeRequestDTO;
-import com.auto_mendes.backend.dto.response.EmployeeResponseDTO;
-import com.auto_mendes.backend.enums.EmployeeType;
+import com.auto_mendes.backend.model.dto.request.EmployeeRequestDTO;
+import com.auto_mendes.backend.model.dto.response.EmployeeResponseDTO;
+import com.auto_mendes.backend.model.enums.EmployeeType;
 
 public interface EmployeeService {
 	EmployeeResponseDTO registerEmployee(EmployeeRequestDTO dto);
