+++ backend/src/test/java/com/auto_mendes/backend/controller/EmployeeControllerTest.java
@@ -19,14 +19,14 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-import com.auto_mendes.backend.dto.request.EmployeeRequestDTO;
-import com.auto_mendes.backend.enums.EmployeeType;
-import com.auto_mendes.backend.persistence.entity.AssistantManager;
-import com.auto_mendes.backend.persistence.entity.Manager;
-import com.auto_mendes.backend.persistence.entity.Saler;
-import com.auto_mendes.backend.persistence.repository.AssistantManagerRepository;
-import com.auto_mendes.backend.persistence.repository.ManagerRepository;
-import com.auto_mendes.backend.persistence.repository.SalerRepository;
+import com.auto_mendes.backend.model.dto.request.EmployeeRequestDTO;
+import com.auto_mendes.backend.model.entity.AssistantManager;
+import com.auto_mendes.backend.model.entity.Manager;
+import com.auto_mendes.backend.model.entity.Saler;
+import com.auto_mendes.backend.model.enums.EmployeeType;
+import com.auto_mendes.backend.repository.AssistantManagerRepository;
+import com.auto_mendes.backend.repository.ManagerRepository;
+import com.auto_mendes.backend.repository.SalerRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
